Made fixes to the code and figures
All figures now has same placement of sensors and emitter for better
comparison between techniques.
JPG version of figures are also added. Zoomed in Minmax is added as JPG
to better show where the estimated location is with respect to the
actual location and the quadrilateral formed.

Some cleanup of unnecessary parameters is done in code such as plotON.
Because there is no cost function computed over region of interest there
is nothing to plot other than placements and estimation location.