Added Performance Simulations
Added figures showing performance evaluation of each localization
technique for different number of sensors and shadow spread values.

Added function PerformanceTest_ShadowSpread that is used to obtain the
figures.

Added function GoldoniTestbed to simulate the experimentation done in
the article by Goldoni.

Changed all method implementations to return estimated emitter location
so that plotting of placements and estimations can be done outside the
function as well.